fix: corrige erros no CI/CD

+++ prisma/seed.ts
@@ -4,17 +4,16 @@ import bcrypt from "bcrypt";
 const prisma = new PrismaClient();
 
 async function main() {
-  console.log("🧹 Limpando o banco de dados...");
+  if (process.env.NODE_ENV === "production") {
+    throw new Error("O seed não pode ser executado em produção.");
+  }
+  console.log("Limpando o banco de dados...");
   
-  // 1. DELEÇÃO NA ORDEM CORRETA
-  // Primeiro deletamos os filhos (Usuários)
   await prisma.user.deleteMany();
-  // Depois deletamos os pais (Roles)
   await prisma.role.deleteMany();
 
-  console.log("🌱 Plantando novos dados iniciais...");
+  console.log("Plantando novos dados iniciais...");
 
-  // 2. CRIAR TODOS OS PAPÉIS (ROLES)
   const adminRole = await prisma.role.create({
     data: { name: "admin" },
   });
@@ -27,10 +26,8 @@ async function main() {
     data: { name: "usuario" },
   });
 
-  // 3. GERAR SENHA DO ADMIN
   const hashPassword = await bcrypt.hash("Admin123!", 10);
 
-  // 4. CRIAR O USUÁRIO ADMIN
   const adminUser = await prisma.user.create({
     data: {
       email: "admin@igarape.com.br",
@@ -39,19 +36,19 @@ async function main() {
       phone: "61999999999",
       password: hashPassword,
       firstAccess: true,
-      roleId: adminRole.id, // Relacionando com o papel 'admin' recém-criado
+      roleId: adminRole.id, 
     },
   });
 
-  console.log("✅ Papéis criados com sucesso: admin, tecnico, usuario.");
-  console.log("✅ Usuário Admin criado com sucesso!");
-  console.log(`📧 Email para teste: ${adminUser.email}`);
-  console.log(`🔑 Senha para teste: Admin123!`);
+  console.log("Papéis criados com sucesso: admin, tecnico, usuario.");
+  console.log("Usuário Admin criado com sucesso!");
+  console.log(`Email para teste: ${adminUser.email}`);
+  console.log(`Senha para teste: Admin123!`);
 }
 
 main()
   .catch((e) => {
-    console.error("❌ Erro ao rodar o seed:", e);
+    console.error("Erro ao rodar o seed:", e);
     process.exit(1);
   })
   .finally(async () => {
